requirements.txt
pandas
lxml
selenium
rapidfuzz
openpyxl
tenacity
nest_asyncio
google-genai
pyyaml
openai
anthropic
